% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uin.R
\name{dob_from_uin}
\alias{dob_from_uin}
\title{Extract date of birth from UIN.}
\usage{
dob_from_uin(uin)
}
\arguments{
\item{uin}{A character variable of the format \code{03121999MAB} i.e. DOB
followed by M/F followed by two initials.}
}
\value{
\code{'1999-12-03'} The DOB in YYYY-MM-DD format.
}
\description{
\code{dob_from_uin} Unique identifiers (UINs) in the MSNZ study have the date
of birth embedded. This function extracts and returns it.
}
