% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectancy.R
\name{expected_year_of_death}
\alias{expected_year_of_death}
\title{Get expected year of death.}
\usage{
expected_year_of_death(
  year_of_birth,
  sex,
  conditional_age,
  conditional_age_default = 39,
  method = "median",
  percentile = "median",
  n_samples = 1
)
}
\arguments{
\item{year_of_birth}{Year of birth as an integer.}

\item{sex}{A character variable containing either \code{female} or
\code{male}.}

\item{conditional_age}{Conditional age as an integer. We use this so that
the expected age of death is conditional on having reached this age,
rather than being calculated from the value at birth. If the conditional
age is \code{NA}, then conditional_age_default is substituted for it.}

\item{conditional_age_default}{The default conditional age to use when
conditional age is \code{NA}. Defaults to 39.}

\item{method}{Specify whether to give the \code{'median'} life
expectancy (the default), or \code{'sample'} from the distribution defined
by the median life expectancy and the variance around this. While the other
arguments can be vectors of multiple values, this is a single value that
applies to everything.}

\item{percentile}{When returning the median life expectancy, specify whether to
return the \code{'median'} estimate of this value (the default), or \code{'5'}
or \code{'95'} for the 5th or 95th percentile. While the other arguments can be
vectors of multiple values, this is a single value that applies to everything.}

\item{n_samples}{When the method is \code{'sample'} set how many samples to
return for each individual. Defaults to 1. While the other arguments can be
vectors of multiple values, this is a single value that applies to everything.}
}
\value{
\code{'YYYY.Y'} The expected year of death (numeric).
}
\description{
\code{expected_year_of_death} Look up the NZ cohort life tables for a person
of the specified year of birth, sex, and age at diagnosis, and return the
expected year of death.
}
