% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uin.R
\name{initials_from_uin}
\alias{initials_from_uin}
\title{Extract initials from UIN.}
\usage{
initials_from_uin(uin)
}
\arguments{
\item{uin}{A character variable of the format \code{03121999MAB} i.e. DOB
followed by M/F followed by two initials.}
}
\value{
\code{'AB'}.
}
\description{
\code{dob_from_uin} Unique identifiers (UINs) in the MSNZ study have the sex
embedded. This function extracts and returns it.
}
