% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msnz.R
\docType{package}
\name{msnz-package}
\alias{msnz}
\alias{msnz-package}
\title{msnz: Functions to Work with New Zealand Multiple Sclerosis Studies}
\description{
This package bundles together useful functions to work with New
    Zealand Multiple Sclerosis epidemiology data.
}
\details{
Package for analysing data from New Zealand Multiple Sclerosis
epidemiology studies.
}
\author{
\strong{Maintainer}: Michael MacAskill \email{michael.macaskill@nzbri.org}

Other contributors:
\itemize{
  \item Daniel Myall \email{daniel.myall@nzbri.org} [contributor]
}

}
\keyword{internal}
