% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uin.R
\name{sex_from_uin}
\alias{sex_from_uin}
\title{Extract sex from UIN.}
\usage{
sex_from_uin(uin, return_as_factor = TRUE)
}
\arguments{
\item{uin}{A character variable of the format \code{03121999MAB} i.e. DOB
followed by M/F followed by two initials.}
}
\value{
\code{'Female', 'Male'} or \code{'F', 'M'}. If
\code{return_as_factor} is \code{TRUE}, then a factor variable is returned,
with labels set as \code{'Female', 'Male'}, and \code{Female} as the
reference level. Otherwise, returns the raw values as \code{'F', 'M'}.
}
\description{
\code{sex_from_uin} Unique identifiers (UINs) in the MSNZ study have the sex
embedded. This function extracts and returns it.
}
