% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectancy.R
\name{expected_year_of_death}
\alias{expected_year_of_death}
\title{Get expected year of death.}
\usage{
expected_year_of_death(
  year_of_birth,
  sex,
  age_at_diagnosis,
  percentile = "median"
)
}
\arguments{
\item{year_of_birth}{Year of birth as an integer.}

\item{sex}{A character variable containing either \code{female} or
\code{male}.}

\item{age_at_diagnosis}{Age at diagnosis as an integer. We use this so that
the expected age of death is conditional on having reached the age of
diagnosis, rather than being calculated from the value at birth. If the age
at diagnosis is \code{NA}, then a median value of 39 is substituted for it.}

\item{percentile}{Specify whether to return the \code{'median'} estimate (the
default), or \code{'5'} or \code{'95'} for the 5th or 95th percentile. While
the other arguments can be vectors of multiple values, this is a single value
that applies to everything.}
}
\value{
\code{'YYYY.Y'} The expected year of death (numeric).
}
\description{
\code{expected_year_of_death} Look up the NZ cohort life tables for a person
of the specified year of birth, sex, and age at diagnosis, and return the
expected year of death.
}
